% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{df_cc_games}
\alias{df_cc_games}
\alias{df_american_checkers}
\alias{df_breakthrough}
\alias{df_crossings}
\alias{df_focus}
\alias{df_four_field_kono}
\alias{df_grasshopper}
\alias{df_jul_gono}
\alias{df_lines_of_action}
\alias{df_turkish_draughts}
\title{Data frames of starting diagrams for various games}
\usage{
df_american_checkers(cell_width = 1)

df_breakthrough(cell_width = 1)

df_crossings(cell_width = 1)

df_focus(cell_width = 1)

df_four_field_kono(cell_width = 1)

df_grasshopper(cell_width = 1)

df_jul_gono(cell_width = 1)

df_lines_of_action(cell_width = 1)

df_turkish_draughts(cell_width = 1)
}
\arguments{
\item{cell_width}{Width of board cell}
}
\description{
\code{tibble} data frames of starting diagrams for various games.
}
\details{
Here is some more information about the various games:
 \describe{
 \item{Breakthrough}{Two-player abstract invented by Dan Troyka in 2000
                     that won the 2001 8x8 Game Design Competition.
                     See \url{https://en.wikipedia.org/wiki/Breakthrough_(board_game)}.}
 \item{(American) Checkers}{A traditional board game also known as \dQuote{(English) Draughts}.
                 See \url{https://en.wikipedia.org/wiki/Draughts}.}
 \item{Crossings}{An abstract invented by Robert Abbot.
                  See \url{https://en.wikipedia.org/wiki/Crossings_(game)}}
 \item{Focus}{An asbtract designed by Sid Sackson.
              See \url{https://en.wikipedia.org/wiki/Focus_(board_game)}.}
 \item{Four Field Kono}{Traditional 2-player Korean abstract.
                        See \url{https://en.wikipedia.org/wiki/Four_Field_Kono}.}
 \item{Grasshopper}{Two-player Halma variant playable on a draughts board.
                    See \url{http://www.cyningstan.com/game/71/grasshopper}.}
 \item{Jul-Gono}{Traditional 2-player Korean abstract.
                 See \url{https://en.wikipedia.org/wiki/Jul-Gonu}.}
 \item{Lines of Action}{An abstract designed by Claude Soucie.
                        See \url{https://en.wikipedia.org/wiki/Lines_of_Action}.}
 \item{Turkish Draughts}{AKA \dQuote{Dama} is a traditional checkers variant played in the Mediterranean.
                         See \url{https://en.wikipedia.org/wiki/Turkish_draughts}.}
}
}
