% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{df_cc_games}
\alias{df_cc_games}
\alias{df_american_checkers}
\alias{df_breakthrough}
\alias{df_four_field_kono}
\alias{df_turkish_draughts}
\title{Data frames of starting diagrams for various games}
\usage{
df_american_checkers(cell_width = 1)

df_breakthrough(cell_width = 1)

df_four_field_kono(cell_width = 1)

df_turkish_draughts(cell_width = 1)
}
\arguments{
\item{cell_width}{Width of board cell}
}
\description{
\code{tibble} data frames of starting diagrams for various games.
}
\details{
Here is some more information about the various games:
 \describe{
 \item{Breakthrough}{Two-player abstract invented by Dan Troyka in 2000
                     that won the 2001 8x8 Game Design Competition.
                     See \url{https://en.wikipedia.org/wiki/Breakthrough_(board_game)}.}
 \item{(American) Checkers}{A traditional board game also known as \dQuote{(English) Draughts}.
                 First adapted to the piecepack by Mark A. Biggar.
                 See \url{http://www.ludism.org/ppwiki/Checkers}.}
 \item{Four Field Kono}{Traditional 2-player Korean abstract adapted to piecepack by Michael Schoessow.
                        See \url{http://www.ludism.org/ppwiki/FourFieldKono}.}
 \item{Turkish Draughts}{AKA \dQuote{Dama} is a traditional checkers variant played in the Mediterranean.
                         See \url{https://en.wikipedia.org/wiki/Turkish_draughts}.}
}
}
